fix(server): enqueue the prebuilt warm-up once the install fully succeeded
Enqueuing from the manifest apply let a synchronous post-install hook or
the registration refresh fail and roll back a fresh install while the
warm-up job was already building lambdas, leaving orphaned functions and
layers. The install service now enqueues after the last step that can
trigger that rollback.
